Создал классы юзера и адреса для representation model, Добавил список исполнителей на главную страницу, и поиск по Городу

+++ src/main/java/com/netcracker/educ/printing/model/repository/CityRepo.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface CityRepo extends JpaRepository<City, Integer> {
     City findAllByTitle(String title);
-    City findAllById(Integer id);
+    City findAllById(Long id);
 }
